refactor(plugin_api): dismantle directory now that all plugins are Lua (#170)
Move `MapApi` (drawing facade) and `PanelAnchor` (layout vocabulary)
out of `src/plugin_api/` and into `src/compositor/`, then delete the
directory. The `plugin_api` name promised a plugin-author API surface,
but every consumer is now `compositor` itself or the Lua bridge —
since the Lua migration retired the in-tree Rust plugin author
prelude. The remaining contents were host-internal types, mis-shelved
under a name that no longer matches reality.

Both files use `git mv` so blame survives. `PanelAnchor::from_str`
had exactly one caller (`lua/component.rs`); inline it there as a
private `parse_panel_anchor` helper rather than carrying it on the
enum. The `Anchor` (screen-space) type stays where `MapApi` lives —
it's not dead at runtime; the Lua bridge wires it via `text_anchored`.

No behaviour change. Compiler enforces correctness via use updates
across `lib.rs`, `compositor/mod.rs`, `lua/component.rs`,
`lua/map_api.rs`, `ui.rs`. Doc references in `CLAUDE.md`, `README.md`,
and `shared/mod.rs` updated to drop `plugin_api`.
