+++ CLAUDE.md
@@ -54,10 +54,9 @@ Key modules:
 - **`src/map/tile/fetch/`**: `TileFetcher` (per-backend trait — "key → bytes"), the generic `FetchLane<F>` (queue / workers / dedup / priority), and the `DiskCachedFetcher<F>` decorator that adds a disk read-through / write-through layer to any inner fetcher. `http.rs` is the only inner backend today; new backends (mbtiles, pmtiles, …) add a `TileFetcher` impl + a branch in `app::build_tile_cache`.
 - **`src/map/tile/decoder.rs`**: Single-thread relay that reads bytes from the fetch lane, calls `decode::decode`, and forwards `DecodedTile`s to the cache. Empty bytes (negative cache from failed fetches) bypass `decode()` and surface as `DecodedTile::empty()`.
 - **`src/app/`**: `App` is the single Receiver and owns the latest `MapFrame` directly (no `UiState` wrapper — built-in chrome lives in plugins now). `src/app/msg.rs` holds the `AppMsg` intent vocabulary (map-level actions nest under `AppMsg::Map(Action)` because `MapState` owns its own vocabulary; other variants sit at the top level); `src/app/mod.rs` holds `App::dispatch` (the thin router) and cross-cutting methods like `apply_theme` / `handle_resize`. Mouse handling lives in `src/app/mouse.rs`. Single side-effect boundary for user-intent state changes. See [docs/design.md](docs/design.md) for the AppMsg-vs-direct-API judgment rules.
-- **`src/compositor/`**: Stack-based focus/modal system (helix-inspired). One primitive: a stack of `Component`s, where the top owns key focus. Replaced the old `FocusManager` + `Plugin` trilogy. Components render map overlays through `Component::paint_on_map`, side panels through `Component::render`, and can emit `AppMsg` via `Window`. Always-on chrome (info, scale_bar, attribution) registers as `overlays` — non-focusable, painted every frame.
+- **`src/compositor/`**: Stack-based focus/modal system (helix-inspired). One primitive: a stack of `Component`s, where the top owns key focus. Replaced the old `FocusManager` + `Plugin` trilogy. Components render map overlays through `Component::paint_on_map`, side panels through `Component::render`, and can emit `AppMsg` via `Window`. Always-on chrome (info, scale_bar, attribution) registers as `overlays` — non-focusable, painted every frame. Submodules `map_api.rs` (drawing facade for `paint_on_map`) and `layout.rs` (`PanelAnchor` vocabulary) live here too — both are crate-internal types consumed by the Lua bridge through `Component`. The earlier `src/plugin_api/` directory that hosted them was a misnamed bucket from the pre-Lua era; it's gone now that every plugin author goes through Lua.
 - **`src/ui.rs`**: Single thin `draw()` entry — no state of its own. Lays out the map area + footer and routes through the compositor for overlays and panels.
 - **`src/palette/`**: `:`-triggered command palette as an ephemeral `Component` (state per-open, discarded on pop). Provider sub-modes (theme picker, forward-geocode search) swap in place via `PaletteAction::SwitchProvider` or are pushed pre-loaded by their key activation. Providers can be sync (`OnEachKey` filter — command, theme) or async (`Debounced` filter, `poll()` to drain results, `is_loading()` for the spinner — search).
-- **`src/plugin_api/`**: Crate-internal primitives the Lua bridge re-uses — `MapApi` (drawing facade) and `PanelAnchor` / `LayoutConfig` (anchor vocabulary). The old plugin-author prelude (`PolledFeed`, `AsyncJob`, `Throttle`, `NominatimClient`, etc.) was retired together with the in-tree Rust plugins; equivalents live in Lua scripts now.
 - **`src/commands/`**: CLI subcommands (currently `snap`). Each subcommand is one file with a `run()` entry point.
 - **`src/keymap.rs`**: Keybinding table + `KeybindingOverrides` (read from `[keymap]` config). Keyboard supports count prefixes for pan (e.g., `5j`); `:` opens the command palette, `/` opens the palette pre-loaded with the search provider.
 - **`src/geo.rs`**: Web Mercator projection math — lon/lat ↔ tile coordinates, distance calculations.
