+++ src/compositor/map_api.rs
